#226 [refact] 프로젝트에서 label 공통 추출부분을 하나의 클래스로 분리하여 호출하는 식으로 리팩토링한다.
