+++ src/main/java/com/dataracy/modules/project/application/service/query/ProjectSearchService.java
@@ -3,19 +3,20 @@ package com.dataracy.modules.project.application.service.query;
 import com.dataracy.modules.common.logging.support.LoggerFactory;
 import com.dataracy.modules.project.application.dto.request.search.FilteringProjectRequest;
 import com.dataracy.modules.project.application.dto.response.search.FilteredProjectResponse;
-import com.dataracy.modules.project.application.dto.response.search.PopularProjectResponse;
+import com.dataracy.modules.project.application.dto.response.read.PopularProjectResponse;
 import com.dataracy.modules.project.application.dto.response.search.RealTimeProjectResponse;
 import com.dataracy.modules.project.application.dto.response.search.SimilarProjectResponse;
 import com.dataracy.modules.project.application.dto.response.support.ProjectLabelMapResponse;
 import com.dataracy.modules.project.application.mapper.search.FilteredProjectDtoMapper;
-import com.dataracy.modules.project.application.mapper.search.PopularProjectDtoMapper;
+import com.dataracy.modules.project.application.mapper.read.PopularProjectDtoMapper;
+import com.dataracy.modules.project.application.port.in.query.extractor.FindProjectLabelMapUseCase;
 import com.dataracy.modules.project.application.port.in.query.search.SearchFilteredProjectsUseCase;
-import com.dataracy.modules.project.application.port.in.query.search.SearchPopularProjectsUseCase;
+import com.dataracy.modules.project.application.port.in.query.read.GetPopularProjectsUseCase;
 import com.dataracy.modules.project.application.port.in.query.search.SearchRealTimeProjectsUseCase;
 import com.dataracy.modules.project.application.port.in.query.search.SearchSimilarProjectsUseCase;
 import com.dataracy.modules.project.application.port.out.query.read.FindProjectPort;
 import com.dataracy.modules.project.application.port.out.query.search.SearchFilteredProjectsPort;
-import com.dataracy.modules.project.application.port.out.query.search.SearchPopularProjectsPort;
+import com.dataracy.modules.project.application.port.out.query.read.GetPopularProjectsPort;
 import com.dataracy.modules.project.application.port.out.query.search.SearchRealTimeProjectsPort;
 import com.dataracy.modules.project.application.port.out.query.search.SearchSimilarProjectsPort;
 import com.dataracy.modules.project.domain.enums.ProjectSortType;
@@ -43,17 +44,14 @@ import java.util.Map;
 public class ProjectSearchService implements
         SearchRealTimeProjectsUseCase,
         SearchSimilarProjectsUseCase,
-        SearchPopularProjectsUseCase,
         SearchFilteredProjectsUseCase
 {
-    private final PopularProjectDtoMapper popularProjectDtoMapper;
     private final FilteredProjectDtoMapper filteredProjectDtoMapper;
 
     private final FindProjectPort findProjectPort;
 
     private final SearchRealTimeProjectsPort searchRealTimeProjectsPort;
     private final SearchSimilarProjectsPort searchSimilarProjectsPort;
-    private final SearchPopularProjectsPort searchPopularProjectsPort;
     private final SearchFilteredProjectsPort searchFilteredProjectsPort;
 
     private final FindUsernameUseCase findUsernameUseCase;
@@ -61,6 +59,7 @@ public class ProjectSearchService implements
     private final GetAnalysisPurposeLabelFromIdUseCase getAnalysisPurposeLabelFromIdUseCase;
     private final GetDataSourceLabelFromIdUseCase getDataSourceLabelFromIdUseCase;
     private final GetAuthorLevelLabelFromIdUseCase getAuthorLevelLabelFromIdUseCase;
+    private final FindProjectLabelMapUseCase findProjectLabelMapUseCase;
 
     /**
      * 주어진 키워드로 실시간 프로젝트를 검색하여 결과 목록을 반환합니다.
@@ -105,59 +104,6 @@ public class ProjectSearchService implements
         return similarProjectResponses;
     }
 
-//    @Override
-//    @Transactional(readOnly = true)
-//    public List<ProjectPopularSearchResponse> findPopularProjects(int size) {
-//        List<Project> savedProjects = projectQueryRepositoryPort.findPopularProjects(size);
-//        List<ProjectPopularSearchResponse> responseDto = savedProjects.stream()
-//                .map(project -> {
-//                    String username = findUsernameUseCase.findUsernameById(project.getUserId());
-//                    String topicLabel = getTopicLabelFromIdUseCase.getLabelById(project.getTopicId());
-//                    String analysisPurposeLabel = getAnalysisPurposeLabelFromIdUseCase.getLabelById(project.getAnalysisPurposeId());
-//                    String dataSourceLabel = getDataSourceLabelFromIdUseCase.getLabelById(project.getDataSourceId());
-//                    String authorLevelLabel = getAuthorLevelLabelFromIdUseCase.getLabelById(project.getAuthorLevelId());
-//
-//                    return popularProjectsDtoMapper.toResponseDto(
-//                            project,
-//                            username,
-//                            topicLabel,
-//                            analysisPurposeLabel,
-//                            dataSourceLabel,
-//                            authorLevelLabel
-//                    );
-//                })
-//                .toList();
-//        return responseDto;
-/**
-     * 지정한 개수만큼 인기 프로젝트를 조회하여, 각 프로젝트에 사용자명과 주제, 분석 목적, 데이터 소스, 저자 레벨 등 라벨 정보를 포함한 응답 리스트를 반환합니다.
-     *
-     * @param size 조회할 인기 프로젝트의 최대 개수
-     * @return 사용자명과 다양한 라벨 정보가 포함된 인기 프로젝트 응답 리스트
-     */
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<PopularProjectResponse> searchPopularProjects(int size) {
-        Instant startTime = LoggerFactory.service().logStart("SearchPopularProjectsUseCase", "인기 프로젝트 목록 조회 서비스 시작");
-
-        List<Project> savedProjects = searchPopularProjectsPort.searchPopularProjects(size);
-
-        ProjectLabelMapResponse labelResponse = labelMapping(savedProjects);
-        List<PopularProjectResponse> popularProjectResponses = savedProjects.stream()
-                .map(project -> popularProjectDtoMapper.toResponseDto(
-                        project,
-                        labelResponse.usernameMap().get(project.getUserId()),
-                        labelResponse.topicLabelMap().get(project.getTopicId()),
-                        labelResponse.analysisPurposeLabelMap().get(project.getAnalysisPurposeId()),
-                        labelResponse.dataSourceLabelMap().get(project.getDataSourceId()),
-                        labelResponse.authorLevelLabelMap().get(project.getAuthorLevelId())
-                ))
-                .toList();
-
-        LoggerFactory.service().logSuccess("SearchPopularProjectsUseCase", "인기 프로젝트 목록 조회 서비스 종료", startTime);
-        return popularProjectResponses;
-    }
-
     /**
      * 필터 조건과 페이지 정보를 기반으로 프로젝트 목록을 조회하고, 각 프로젝트에 사용자명 및 다양한 라벨 정보를 매핑하여 페이지 형태로 반환합니다.
      *
@@ -174,7 +120,7 @@ public class ProjectSearchService implements
                 : null;
         Page<Project> savedProjects = searchFilteredProjectsPort.searchByFilters(request, pageable, sortType);
 
-        ProjectLabelMapResponse labelResponse = labelMapping(savedProjects.getContent());
+        ProjectLabelMapResponse labelResponse = findProjectLabelMapUseCase.labelMapping(savedProjects.getContent());
         Page<FilteredProjectResponse> filteredProjectResponses = savedProjects.map(project -> {
             // 자식 프로젝트들의 userId 수집
             List<Long> childUserIds = project.getChildProjects().stream()
@@ -198,27 +144,5 @@ public class ProjectSearchService implements
 
         LoggerFactory.service().logSuccess("SearchFilteredProjectsUseCase", "인기 프로젝트 목록 조회 서비스 종료 keyword=" + request.keyword(), startTime);
         return filteredProjectResponses;
-    }
-
-    /**
-     * 프로젝트 컬렉션에서 사용자, 토픽, 분석 목적, 데이터 소스, 저자 레벨의 ID를 추출하여 각 ID에 해당하는 사용자명과 레이블 정보를 일괄 조회합니다.
-     *
-     * @param savedProjects 사용자명 및 레이블 매핑 정보를 조회할 프로젝트 컬렉션
-     * @return 각 ID에 대한 사용자명과 레이블 매핑 정보를 포함하는 ProjectLabelMapResponse 객체
-     */
-    private ProjectLabelMapResponse labelMapping(Collection<Project> savedProjects) {
-        List<Long> userIds = savedProjects.stream().map(Project::getUserId).toList();
-        List<Long> topicIds = savedProjects.stream().map(Project::getTopicId).toList();
-        List<Long> analysisPurposeIds = savedProjects.stream().map(Project::getAnalysisPurposeId).toList();
-        List<Long> dataSourceIds = savedProjects.stream().map(Project::getDataSourceId).toList();
-        List<Long> authorLevelIds = savedProjects.stream().map(Project::getAuthorLevelId).toList();
-
-        return new ProjectLabelMapResponse(
-                findUsernameUseCase.findUsernamesByIds(userIds),
-                getTopicLabelFromIdUseCase.getLabelsByIds(topicIds),
-                getAnalysisPurposeLabelFromIdUseCase.getLabelsByIds(analysisPurposeIds),
-                getDataSourceLabelFromIdUseCase.getLabelsByIds(dataSourceIds),
-                getAuthorLevelLabelFromIdUseCase.getLabelsByIds(authorLevelIds)
-        );
     }
 }
